feat(Makefile): add package-keycloak target to build and push keycloak docker image feat(docker-compose-keycloak.yml): replace i2-keycloak image with connect-admin-keycloak feat(docker-compose-keycloak.yml): expose postgres port for external access feat(keycloak/Dockerfile): add new Dockerfile for keycloak with custom build steps feat(keycloak/package.json): update keycloak-js version and license fix(keycloak/index.html): correct typo in meta description feat(keycloak/public/env-config.js): add new env-config.js file for environment variables feat(.env_dev): add CONNECT_ADMIN_KEYCLOAK_VERSION variable for keycloak image version
These changes were made to add support for a custom keycloak docker image and expose the postgres port for external access. The keycloak-js version and license were updated in package.json. A typo in the meta description of index.html was corrected. A new env-config.js file was added for environment variables