chore(docker): set NODE_OPTIONS environment variable to increase max old space size to 4096 chore(package.json): update dependencies for @smartb domains to version 0.15.0-experimental.1695388358212 refactor(model.ts): remove unused import and roles property from ApiKeyDTO interface
The Dockerfile was modified to set the NODE_OPTIONS environment variable to increase the maximum old space size to 4096. This is done to allocate more memory for the Node.js process running in the Docker container.

In the package.json file, the dependencies for the @smartb domains (@smartb/privilege-domain, @smartb/apikey-domain, @smartb/user-domain, @smartb/organization-domain) were updated to version 0.15.0-experimental.1695388358212. This is done to incorporate the latest changes and improvements in the domains.

In the model.ts file, an unused import for "components" was removed, and the "roles" property was removed from the ApiKeyDTO interface. This is done to clean up the code and remove unused code and dependencies.